board/requirements.txt
Adafruit-Blinka==8.65.0
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-dht==4.0.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.1
Adafruit-PlatformDetect==3.83.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
board==1.0
lgpio==0.2.2.0
paho-mqtt==2.1.0
pyftdi==0.57.1
pyserial==3.5
pyusb==1.3.1
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.15.0
pip==23.0.1
setuptools